EntityMetaData moved to dto package
