+++ src/main/java/org/foxminded/springcourse/consoleapp/manager/EntityMetaDataManager.java
@@ -1,8 +1,8 @@
 package org.foxminded.springcourse.consoleapp.manager;
 
+import org.foxminded.springcourse.consoleapp.dto.EntityMetaData;
+import org.foxminded.springcourse.consoleapp.dto.EntityMetaDataCache;
 import org.foxminded.springcourse.consoleapp.service.EntityMetaDataExtractor;
-import org.foxminded.springcourse.consoleapp.model.EntityMetaData;
-import org.foxminded.springcourse.consoleapp.model.EntityMetaDataCache;
 import org.springframework.stereotype.Component;
 
 @Component
