+++ src/main/java/org/foxminded/springcourse/consoleapp/service/EntityDataMapper.java
@@ -2,9 +2,9 @@ package org.foxminded.springcourse.consoleapp.service;
 
 import org.foxminded.springcourse.consoleapp.annotation.Column;
 import org.foxminded.springcourse.consoleapp.annotation.Id;
+import org.foxminded.springcourse.consoleapp.dto.EntityMetaData;
 import org.foxminded.springcourse.consoleapp.exception.EntityDataMapperException;
 import org.foxminded.springcourse.consoleapp.manager.EntityMetaDataManager;
-import org.foxminded.springcourse.consoleapp.model.EntityMetaData;
 import org.springframework.stereotype.Service;
 import org.springframework.util.ReflectionUtils;
