+++ src/test/java/org/foxminded/springcourse/consoleapp/dao/CourseDaoTest.java
@@ -1,9 +1,9 @@
 package org.foxminded.springcourse.consoleapp.dao;
 
+import org.foxminded.springcourse.consoleapp.dto.EntityMetaDataCache;
 import org.foxminded.springcourse.consoleapp.manager.EntityMetaDataManager;
 import org.foxminded.springcourse.consoleapp.model.ConnectionConfig;
 import org.foxminded.springcourse.consoleapp.model.Course;
-import org.foxminded.springcourse.consoleapp.model.EntityMetaDataCache;
 import org.foxminded.springcourse.consoleapp.service.EntityDataMapper;
 import org.foxminded.springcourse.consoleapp.service.EntityMetaDataExtractor;
 import org.junit.jupiter.api.AfterEach;
